feat: make dyn bitcoin rpc available in ServerModuleInitArgs
