+++ fedimint-core/src/module/mod.rs
@@ -31,6 +31,7 @@ use tracing::Instrument;
 // TODO: Make this module public and remove the wildcard `pub use` below
 mod version;
 pub use self::version::*;
+use crate::bitcoin_rpc::DynBitcoindRpc;
 use crate::config::{
     ClientModuleConfig, ConfigGenModuleParams, DkgPeerMsg, ModuleInitParams, ServerModuleConfig,
     ServerModuleConsensusConfig,
@@ -500,6 +501,7 @@ pub trait IServerModuleInit: IDynCommonModuleInit {
         db: Database,
         task_group: &TaskGroup,
         our_peer_id: PeerId,
+        dyn_bitcoin_rpc: DynBitcoindRpc,
     ) -> anyhow::Result<DynServerModule>;
 
     fn validate_params(&self, params: &ConfigGenModuleParams) -> anyhow::Result<()>;
@@ -580,6 +582,7 @@ where
     task_group: TaskGroup,
     our_peer_id: PeerId,
     num_peers: NumPeers,
+    dyn_bitcoin_rpc: DynBitcoindRpc,
     // ClientModuleInitArgs needs a bound because sometimes we need
     // to pass associated-types data, so let's just put it here right away
     _marker: marker::PhantomData<S>,
@@ -607,6 +610,10 @@ where
 
     pub fn our_peer_id(&self) -> PeerId {
         self.our_peer_id
+    }
+
+    pub fn dyn_bitcoin_rpc(&self) -> DynBitcoindRpc {
+        self.dyn_bitcoin_rpc.clone()
     }
 }
 /// Module Generation trait with associated types
@@ -694,6 +701,7 @@ where
         db: Database,
         task_group: &TaskGroup,
         our_peer_id: PeerId,
+        dyn_bitcoin_rpc: DynBitcoindRpc,
     ) -> anyhow::Result<DynServerModule> {
         <Self as ServerModuleInit>::init(
             self,
@@ -703,6 +711,7 @@ where
                 db,
                 task_group: task_group.clone(),
                 our_peer_id,
+                dyn_bitcoin_rpc,
                 _marker: PhantomData,
             },
         )
