+++ fedimintd/src/fedimintd.rs
@@ -420,6 +420,7 @@ impl Fedimintd {
                 self.server_gens,
                 self.server_gen_params,
                 self.code_version_str,
+                self.bitcoind_rpc,
             )
             .await
             {
@@ -536,6 +537,7 @@ async fn run(
     module_inits: ServerModuleInitRegistry,
     module_inits_params: ServerModuleConfigGenParamsRegistry,
     code_version_str: String,
+    bitcoin_rpc: BitcoinRpcConfig,
 ) -> anyhow::Result<()> {
     if let Some(socket_addr) = opts.bind_metrics_api.as_ref() {
         task_group.spawn_cancellable("metrics-server", {
@@ -582,6 +584,7 @@ async fn run(
         code_version_str,
         &module_inits,
         task_group.clone(),
+        bitcoin_rpc,
     )
     .await?;
